add producer and consumer model

+++ client/CoinMiner/app/src/main/java/com/coding/coinminer/data/Model.kt
@@ -7,6 +7,7 @@ package com.coding.coinminer.data
 
 import com.coding.coinminer.data.Model.MiningData
 import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 
 object Model {
@@ -22,7 +23,7 @@ object Model {
         // Volatile fields provide memory visibility
         // guarantee that the value that is being read, comes from the main memory and not the cpu-cache
         @Volatile
-        var Nonce: AtomicInteger = AtomicInteger()
+        var Nonce: AtomicLong = AtomicLong()
 
     }
 
@@ -33,7 +34,7 @@ object Model {
         MiningData.merkleRoot = i.merkleRoot
         MiningData.timestamp = i.timestamp
         MiningData.difficultyTarget = i.difficultyTarget
-        MiningData.Nonce = AtomicInteger(i.Nonce.toInt())
+        MiningData.Nonce = AtomicLong(i.Nonce.toLong())
 
     }
